Split runner structured delete tests
